add stg_dim_delivery_method

+++ models/staging/stg_dim_delivery_method.sql
@@ -13,7 +13,7 @@ WITH stg_dim_delivery_method__source AS (
 
 , stg_dim_delivery_method__cast AS (
   SELECT 
-    CAST(delivery_method_id AS INT64) AS delivery_method_id
+    CAST(delivery_method_id AS INTEGER) AS delivery_method_id
     , CAST(delivery_method_name AS STRING) AS delivery_method_name
   FROM stg_dim_delivery_method__rename
 )
